requirements.txt
requests
pillow
ultralytics
opencv-python-headless